[Test] `deque.test.ts` and `Arrays.fromSet` API
